+++ src/_views/status-bar.view.ts
@@ -9,8 +9,19 @@ export class StatusBarView {
     private static _items: vscode.StatusBarItem[] = [];
     private static _context: vscode.ExtensionContext;
 
+    private static _cancelRunningItem: vscode.StatusBarItem;
+
     static injectCtx(context: vscode.ExtensionContext) : void {
         StatusBarView._context = context;
+
+        StatusBarView._cancelRunningItem =
+            vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 0);
+        StatusBarView._cancelRunningItem.text = 'Cancel running';
+        StatusBarView._cancelRunningItem.command = {
+            title: 'Cancel running',
+            command: CommandService.prepareCommand(CommandService.cancelRunningCmd),
+        };
+        this._context.subscriptions.push(StatusBarView._cancelRunningItem);
     }
 
     static refresh() {
@@ -53,5 +64,15 @@ export class StatusBarView {
         for (let item of this._items) {
             item.hide();
         }
+    }
+
+    static showCancelRunningItem() : void {
+        StatusBarView.hide();
+        StatusBarView._cancelRunningItem.show();
+    }
+
+    static hideCancelRunningItem() : void {
+        StatusBarView.show();
+        StatusBarView._cancelRunningItem.hide();
     }
 }