+++ src/extension.ts
@@ -33,6 +33,10 @@ import { HideEnvOnStatusBarCommand } from './_commands/hide-env-on-status-bar.co
 import { ShowEnvOnStatusBarCommand } from './_commands/show-env-on-status-bar.command';
 import { WikiView } from './_views/wiki.view';
 import { OpenWikiURLCommand } from './_commands/open-wiki-url.command';
+import { CancelRunningCommand } from './_commands/cancel_running.command';
+import { GoNotebook } from './_notebooks/go.notebook';
+
+let _notebook: GoNotebook;
 
 export function activate(context: vscode.ExtensionContext) {
     // register views.
@@ -91,6 +95,8 @@ export function activate(context: vscode.ExtensionContext) {
         context, CommandService.showOnStatusBarEnvCmd, new ShowEnvOnStatusBarCommand());
     CommandService.registerCommand(
         context, CommandService.openWikiURLCmd, new OpenWikiURLCommand());
+    CommandService.registerCommand(
+        context, CommandService.cancelRunningCmd, new CancelRunningCommand());
 
     // set global listeners.
     const statusBarVisibilityListener = (doc: vscode.TextDocument) => {
@@ -108,6 +114,10 @@ export function activate(context: vscode.ExtensionContext) {
         }
         return statusBarVisibilityListener(e.document);
     });
+
+    _notebook = new GoNotebook(context);
 }
 
-export function deactivate() {}
+export function deactivate() {
+    _notebook?.dispose();
+}